requirements.txt
ultralytics>=8.1.0
deep-sort-realtime>=1.3.2
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
